refactor: refactor video folder
