+++ BackendServer/src/video-insert/video-insert.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { VideoInsertService } from './video-insert.service';
 import { VideoInsertController } from './video-insert.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Video } from '../entities/video.entity';
+import { Video } from 'src/videos/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Video])],
