+++ BackendServer/src/videos/videos.module.ts
@@ -2,10 +2,11 @@ import { Module } from '@nestjs/common';
 import { VideosController } from './videos.controller';
 import { VideosService } from './videos.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Video } from '../entities/video.entity';
+import { Video } from './entities';
+import { UserModule } from 'src/auth/user/user.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Video])],
+  imports: [TypeOrmModule.forFeature([Video]), UserModule],
   controllers: [VideosController],
   providers: [VideosService],
 })
